feat: add auth mode and refactor frontend
