+++ worker/src/worker.js
@@ -8,6 +8,13 @@ import { email } from './email';
 const app = new Hono()
 app.use('/*', cors());
 app.use('/api/*', async (c, next) => {
+	// check header x-custom-auth
+	if (c.env.PASSWORDS && c.env.PASSWORDS.length > 0) {
+		const auth = c.req.headers.get("x-custom-auth");
+		if (!auth || !c.env.PASSWORDS.includes(auth)) {
+			return c.text("Need Password", 401)
+		}
+	}
 	if (c.req.path.startsWith("/api/new_address")) {
 		await next();
 		return;
